refactor: move some classes into two new packages
