+++ src/main/java/hazell/Storage.java
@@ -1,5 +1,7 @@
 package hazell;
 
+import hazell.entities.Task;
+
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
